AddPersonnelFormDropdown - moved onSelect into own function using react.callback. renamed useBasePositionsContext to useBasePosition. Changed use of useBasePosition to be direct instead of through appContext
