+++ src/frontend/src/appContext.ts
@@ -2,12 +2,10 @@ import ApiClient from './api/ApiClient';
 import { createContext, useContext, Dispatch } from 'react';
 import { AppState } from './reducers/appReducer';
 import { CollectionAction } from './reducers/utils';
-import { useBasePositionsContext } from './hooks/useBasePositionsContext';
 
 export interface IAppContext {
     apiClient: ApiClient;
     appState: AppState;
-    useBasePositions: useBasePositionsContext;
     dispatchAppAction: Dispatch<CollectionAction<AppState, keyof AppState>>;
 }
